Re #1700 renamed bin_io_handler.hpp into bin_io_handler.h to maintain project consistency
