+++ _LowLevelCode/cpp/mex_bin_plugin/bin_io_handler.cpp
@@ -1,4 +1,4 @@
-#include "bin_io_handler.hpp"
+#include "bin_io_handler.h"
 
 void bin_io_handler::init(const fileParameters& fpar) {
