+++ _LowLevelCode/cpp/test/mex_bin_plugin.tests/mex_bin_plugin.test.cpp
@@ -1,4 +1,4 @@
-#include "mex_bin_plugin/bin_io_handler.hpp"
+#include "mex_bin_plugin/bin_io_handler.h"
 #include "utility/environment.h"
 #include <gtest/gtest.h>
